Appointment Agent/crew agent/requirements.txt
streamlit
crewai
groq
google-generativeai
python-dotenv
